Moved in-mem repo to src
