+++ apps/admin-x-settings/src/main.tsx
@@ -2,6 +2,7 @@ import './styles/demo.css';
 import App from './App.tsx';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
+import {DefaultHeaderTypes} from './utils/unsplash/UnsplashTypes.ts';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
     <React.StrictMode>
@@ -30,6 +31,7 @@ ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
                 image: 'assets/img/themes/Edition.png'
             }]}
             toggleFeatureFlag={() => {}}
+            unsplashConfig={{} as DefaultHeaderTypes}
             zapierTemplates={[]}
         />
     </React.StrictMode>
